refactor(dh/metering-point): move request-service-action to end-of-supply
The action helper is specific to end-of-supply, not shared across
business reasons.

Refs #1386
